+++ DatadogSessionReplay/Sources/Recorder/RecordingCoordinator.swift
@@ -9,12 +9,15 @@ import Foundation
 import DatadogInternal
 
 /// Object is responsible for getting the RUM context, randomising the sampling rate,
-/// managing the recording state, starting/stopping the recording scheduler as needed,
+/// managing the recording state, starting/stopping the recording,
 /// and propagating `has_replay` to other features.
-
+///
+/// The `RecordingCoordinator` is responsible for orchestrating the process of capturing
+/// snapshots on the main thread.
+///
+/// Subsequent requests to capture snapshots are throttled to avoid performance issues.
 internal class RecordingCoordinator {
     let recorder: Recording
-    let scheduler: Scheduler
     let sampler: Sampler
     let textAndInputPrivacy: TextAndInputPrivacyLevel
     let imagePrivacy: ImagePrivacyLevel
@@ -22,19 +25,26 @@ internal class RecordingCoordinator {
     let srContextPublisher: SRContextPublisher
 
     private var currentRUMContext: RUMContext? = nil
-    private var isSampled = false
+    private var isSampled: Bool = false
 
-    /// `recordingEnabled` is used to track when the user 
+    /// `recordingEnabled` is used to track when the user
     /// has enabled or disabled the recording for Session Replay.
     private var recordingEnabled = false
+
+    private var isRecording = false
 
     /// Sends telemetry through sdk core.
     private let telemetry: Telemetry
     /// The sampling rate for internal telemetry of method calls.
     private let methodCallTelemetrySamplingRate: Float
 
+    private let dateProvider: DateProvider
+
+    private var recordingTrigger: RecordingTriggering
+
+    private let queue: Queue
+
     init(
-        scheduler: Scheduler,
         textAndInputPrivacy: TextAndInputPrivacyLevel,
         imagePrivacy: ImagePrivacyLevel,
         touchPrivacy: TouchPrivacyLevel,
@@ -43,35 +53,34 @@ internal class RecordingCoordinator {
         recorder: Recording,
         sampler: Sampler,
         telemetry: Telemetry,
-        startRecordingImmediately: Bool,
-        methodCallTelemetrySamplingRate: Float = 0.1
-    ) {
+        recordingTrigger: RecordingTriggering,
+        methodCallTelemetrySamplingRate: Float = 0.1,
+        dateProvider: DateProvider = SystemDateProvider(),
+        queue: Queue = MainAsyncQueue()
+    ) throws {
         self.recorder = recorder
-        self.scheduler = scheduler
         self.sampler = sampler
         self.textAndInputPrivacy = textAndInputPrivacy
         self.imagePrivacy = imagePrivacy
         self.touchPrivacy = touchPrivacy
         self.srContextPublisher = srContextPublisher
         self.telemetry = telemetry
+        self.recordingTrigger = recordingTrigger
         self.methodCallTelemetrySamplingRate = methodCallTelemetrySamplingRate
+        self.dateProvider = dateProvider
+        self.queue = queue
 
         srContextPublisher.setHasReplay(false)
 
-        scheduler.schedule { [weak self] in self?.captureNextRecord() }
-
-        // Start recording immediately if specified.
-        if startRecordingImmediately {
-            startRecording()
+        // Observe changes in the RUM context on the main thread.
+        rumContextObserver.observe(on: queue) { [weak self] in
+            self?.onRUMContextChanged(rumContext: $0)
         }
-
-        // Observe changes in the RUM context.
-        rumContextObserver.observe(on: scheduler.queue) { [weak self] in self?.onRUMContextChanged(rumContext: $0) }
     }
 
     /// Enables recording based on user request.
     func startRecording() {
-        scheduler.queue.run { [weak self] in
+        queue.run { [weak self] in
             self?.recordingEnabled = true
             self?.evaluateRecordingConditions()
         }
@@ -79,23 +88,27 @@ internal class RecordingCoordinator {
 
     /// Disables recording based on user request.
     func stopRecording() {
-        scheduler.queue.run { [weak self] in
+        queue.run { [weak self] in
             self?.recordingEnabled = false
             self?.evaluateRecordingConditions()
         }
     }
 
-    // MARK: Private
-
-    /// Evaluates whether recording should start or stop based on user request and sampling.
     private func evaluateRecordingConditions() {
-       if recordingEnabled && isSampled {
-           scheduler.start()
-       } else {
-           scheduler.stop()
-       }
-       updateHasReplay()
-   }
+        if !isRecording && recordingEnabled && isSampled && currentRUMContext != nil {
+            isRecording = true
+            recordingTrigger.startWatchingTriggers { [weak self] in
+                self?.didTrigger()
+            }
+            updateHasReplay()
+        } else if isRecording && (!recordingEnabled || !isSampled) {
+            isRecording = false
+            recordingTrigger.stopWatchingTriggers()
+            updateHasReplay()
+        }
+    }
+
+    // MARK: Private
 
     private func onRUMContextChanged(rumContext: RUMContext?) {
         if currentRUMContext?.sessionID != rumContext?.sessionID || currentRUMContext == nil {
@@ -115,9 +128,22 @@ internal class RecordingCoordinator {
         srContextPublisher.setHasReplay(hasReplay)
     }
 
-    /// Captures the next recording if conditions are met.
+    private var lastTriggerDate: Date?
+
+    private var shouldSkipTrigger: Bool {
+        return dateProvider.now.timeIntervalSince(lastTriggerDate ?? .distantPast) < Constants.throttlingRate
+    }
+
+    private func didTrigger() {
+        guard shouldSkipTrigger == false else {
+            return
+        }
+        lastTriggerDate = dateProvider.now
+        captureNextRecord()
+    }
+
     private func captureNextRecord() {
-        /// We don't capture any snapshots if the RUM context has no view ID.
+        // We don't capture any snapshots if the RUM context has no view ID.
         guard let rumContext = currentRUMContext,
               let viewID = rumContext.viewID else {
             return
@@ -141,7 +167,9 @@ internal class RecordingCoordinator {
 
         var isSuccessful = false
         do {
-            try objc_rethrow { try recorder.captureNextRecord(recorderContext) }
+            try objc_rethrow { [weak self] in
+                try self?.recorder.captureNextRecord(recorderContext)
+            }
             isSuccessful = true
         } catch let objc as ObjcException {
             telemetry.error("[SR] Failed to take snapshot due to Objective-C runtime exception", error: objc.error)
@@ -155,6 +183,10 @@ internal class RecordingCoordinator {
         }
 
         telemetry.stopMethodCalled(methodCalledTrace, isSuccessful: isSuccessful)
+    }
+
+    private enum Constants {
+        static let throttlingRate: TimeInterval = 0.1 // 100ms
     }
 
     private enum MethodCallConstants {